Reorgenize Code
